added carousel view, added delete to Posts

+++ boards/views.py
@@ -196,9 +196,9 @@ def delete_post(request, board_pk, topic_pk, post_pk):
     post = get_object_or_404(
         Post, 
         topic__pk=topic_pk,
-        created_by=request.user,
         pk=post_pk
     )
-    post.delete()
+    if post.created_by == request.user:
+        post.delete()
 
     return redirect('boards:topic_posts', board_pk=post.topic.board.pk, topic_pk=post.topic.pk)
